run test suite
normally the easiest would be to install from the github
tarball and simply specifiy the `tests` folder in the test
section as under `source_files`, however it is very painful
to build aiohttp fully from source. Not least:
 * the submodule for the vendored llhttp needs to be included
 * need to cythonize, but `make cythonize` tries to reinstall
 * the C files refer to _build_ files of the vendored llhttp
 * building llhttp cannot be done (via CMake) without patching
 * building llhttp is only necessary for a few {.c,.h} files,
   and we need to avoid installing something into $PREFIX

So just have a parallel download of the sources to use for the tests.